# Trie node memory optimization fixed