--- a/src/Nethermind/Nethermind.Benchmark/Store/TrieNodeBenchmarks.cs
+++ b/src/Nethermind/Nethermind.Benchmark/Store/TrieNodeBenchmarks.cs
@@ -1,0 +1,70 @@
+//  Copyright (c) 2018 Demerzel Solutions Limited
+//  This file is part of the Nethermind library.
+//
+//  The Nethermind library is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Lesser General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  The Nethermind library is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+//  GNU Lesser General Public License for more details.
+//
+//  You should have received a copy of the GNU Lesser General Public License
+//  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
+
+using System;
+using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
+using Nethermind.Core;
+using Nethermind.Core.Crypto;
+using Nethermind.Core.Test.Builders;
+using Nethermind.State;
+using Nethermind.Db.Blooms;
+using Nethermind.Trie;
+
+namespace Nethermind.Benchmarks.Store
+{
+    [MemoryDiagnoser]
+    public class TrieNodeBenchmarks
+    {
+        [Benchmark]
+        public void PropertiesSet()
+        {
+            TrieNode node = new(NodeType.Extension);
+
+            node.LastSeen = 1;
+            node.IsPersisted = true;
+            node.IsBoundaryProofNode = true;
+
+            node.LastSeen = 0xFF;
+            node.IsPersisted = false;
+            node.IsBoundaryProofNode = false;
+
+            node.LastSeen = 0xFFFF;
+            node.IsPersisted = true;
+            node.IsBoundaryProofNode = true;
+
+            node.LastSeen = 0xFFFFFF;
+            node.IsPersisted = false;
+            node.IsBoundaryProofNode = false;
+
+            node.LastSeen = 0xFFFFFFFF;
+            node.IsPersisted = true;
+            node.IsBoundaryProofNode = true;
+
+            node.LastSeen = 0xFFFFFFFFFF;
+            node.IsPersisted = false;
+            node.IsBoundaryProofNode = false;
+
+            node.LastSeen = 0xFF_FFFF_FFFF;
+            node.IsPersisted = true;
+            node.IsBoundaryProofNode = true;
+
+            node.LastSeen = 0xFFFF_FFFF_FFFF;
+            node.IsPersisted = false;
+            node.IsBoundaryProofNode = false;
+        }
+    }
+}